+++ CHANGES.rst
@@ -10,6 +10,7 @@ Changes
 * Renamed ``irc.client.IRC`` to ``irc.client.Manifold`` to provide a clearer
   name for that object. Clients supporting 8.6 and later can use the
   ``Manifold`` name. Latest clients must use the ``Manifold`` name.
+* Removed unused ``irc.client.mask_matches`` function.
 
 9.0
 ===